- CHG: Included ics-database-commons.jar for database interactions. - CHG: Rewrote code to use the new library. - ADD: Added basic tests for the API. - ADD: Added a PUT method for datasources to check authentication. - ADD: Added a basic master username and password that can be used to insert data via the API. This may be changed to a more refined system in the future. - ADD: Added the main database template MySQL script, as it wasn't included before.

+++ src/jhi/buntata/resource/BuntataAttribute.java
@@ -20,6 +20,8 @@ import com.fasterxml.jackson.annotation.*;
 
 import java.util.*;
 
+import jhi.database.shared.util.*;
+
 /**
  * @author Sebastian Raubach
  */
@@ -35,7 +37,7 @@ public class BuntataAttribute extends DatabaseObject
 	{
 	}
 
-	public BuntataAttribute(int id, Date createdOn, Date updatedOn)
+	public BuntataAttribute(Long id, Date createdOn, Date updatedOn)
 	{
 		super(id, createdOn, updatedOn);
 	}
